task 6

+++ 0x00-hello_world/6-size.c
@@ -3,16 +3,16 @@
 /**
  * main - Entry Point
  *
- * Description: using sizeof to print the size of various types
+ * Description: using sizeof to print the size
  *
  * Return:Always 0 (success)
  */
 int main(void)
 {
-	printf("size of a char: %1u byte(s)\n", (unsigned long)sizeof(char));
-	printf("size of a int: %1u byte(s)\n", (unsigned long)sizeof(int));
-	printf("size of a long int: %1u byte(s)\n", (unsigned long)sizeof(long int));
-	printf("size of a long long int: %1u byte(s)\n", (unsigned long)sizeof(long long int));
-	printf("size of a float: %1u byte(s)\n", (unsigned long)sizeof(float));
+	printf("size of a char: %d byte(s)\n", sizeof(char));
+	printf("size of a int: %d byte(s)\n", sizeof(int));
+	printf("size of a long int: %d byte(s)\n", sizeof(long int));
+	printf("size of a long long int: %d byte(s)\n", sizeof(long long int));
+	printf("size of a float: %d byte(s)\n", sizeof(float));
 	return (0);
 }
